Break dependency between PropertiesUtil and the bulk of the compiler
